archive/agentic_evaluation/requirements.txt
"""
Requirements for Gemini-powered medical evaluation
"""

# Core dependencies
google-cloud-aiplatform
vertexai

# Medical analysis
opencv-python
mediapipe
numpy
Pillow

# Data handling
pandas
json5
# Agent Garden (conceptual - adjust based on actual package)
# agent-garden

# Optional: for enhanced functionality
matplotlib
seaborn
plotly